backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic>=2.5.0
sqlalchemy>=2.0.33
psycopg2-binary>=2.9.11
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[argon2]>=1.7.4
email-validator==2.1.0
